+++ rtvi-client-js/src/clients/index.ts
@@ -1,5 +1,16 @@
-import { BotReadyData, PipecatMetrics, RTVIMessage } from "../messages";
-import { Participant, TransportState } from "../transport";
+import EventEmitter from "events";
+import TypedEmitter from "typed-emitter";
+
+import { RTVIActionRequestData } from "../actions";
+import { RTVIEvents } from "../events";
+import { RTVIClientHelper } from "../helpers";
+import {
+  BotReadyData,
+  PipecatMetricsData,
+  RTVIMessage,
+  RTVIMessageActionResponse,
+} from "../messages";
+import { Participant, Transport, TransportState } from "../transport";
 
 export type RTVIEventCallbacks = Partial<{
   onGenericMessage: (data: unknown) => void;
@@ -16,33 +27,27 @@ export type RTVIEventCallbacks = Partial<{
   onBotDisconnected: (participant: Participant) => void;
   onParticipantJoined: (participant: Participant) => void;
   onParticipantLeft: (participant: Participant) => void;
-  onMetrics: (data: PipecatMetrics) => void;
+  onMetrics: (data: PipecatMetricsData) => void;
 }>;
 
 export interface RTVIClientOptions {
   /**
-   * Base URL for auth handlers and transport services
-   *
-   * Defaults to a POST request with a the config object as the body
-   */
-  baseUrl: string;
-
-  /**
    * Parameters passed as JSON stringified body params to the baseUrl
    */
-  startParams?: Partial<{
-    config?: RTVIClientConfigOption[];
-  }>;
-
-  /**
-   * HTTP headers to be send with the POST request to baseUrl
-   */
-  startHeaders?: { [key: string]: string };
+  params: RTVIClientParams;
 
   /**
    * Optional callback methods for rtvi events
    */
   callbacks?: RTVIEventCallbacks;
+
+  /**
+   * Set transport class for media streaming
+   */
+  transport?: new (
+    options: RTVIClientOptions,
+    onMessage: (ev: RTVIMessage) => void
+  ) => Transport;
 
   /**
    * Handshake timeout
@@ -55,14 +60,13 @@ export interface RTVIClientOptions {
   /**
    * Custom start method handler for retrieving auth bundle for transport
    * @param baseUrl
-   * @param startParams
+   * @param params
    * @param timeout
    * @param abortController
    * @returns Promise<void>
    */
   customAuthHandler?: (
-    baseUrl: string,
-    startParams: object,
+    params: RTVIClientParams,
     timeout: ReturnType<typeof setTimeout> | undefined,
     abortController: AbortController
   ) => Promise<void>;
@@ -70,14 +74,22 @@ export interface RTVIClientOptions {
   // ----- deprecated options
 
   /**
+   * Base URL for auth handlers and transport services
+   *
+   * Defaults to a POST request with a the config object as the body
+   * @deprecated Use params.baseUrl instead
+   */
+  baseUrl?: string;
+
+  /**
    * Service key value pairs (e.g. {llm: "openai"} )
-   * @deprecated Use startParams.services instead
+   * @deprecated Use params.services instead
    */
   services?: VoiceClientServices;
 
   /**
    * Service configuration options for services and further customization
-   * @deprecated Use startParams.config instead
+   * @deprecated Use params.config instead
    */
   config?: VoiceClientConfigOption[];
 
@@ -89,7 +101,7 @@ export interface RTVIClientOptions {
 
   /**
    * Custom request parameters to send with the POST request to baseUrl
-   * @deprecated Use startParams instead
+   * @deprecated Use params instead
    */
   customBodyParams?: object;
 }
@@ -103,6 +115,67 @@ export type RTVIClientConfigOption = {
   service: string;
   options: ConfigOption[];
 };
+
+export type RTVIClientParams = {
+  baseUrl: URL;
+} & Partial<{
+  headers?: Headers;
+  config?: RTVIClientConfigOption[];
+}>;
+
+// ----- Abstract base client
+
+export abstract class RTVIClient extends (EventEmitter as unknown as new () => TypedEmitter<RTVIEvents>) {
+  abstract params: RTVIClientParams;
+  protected abstract _transport: Transport;
+
+  abstract get connected(): boolean;
+  abstract connect(): Promise<unknown>;
+  abstract disconnect(): Promise<void>;
+
+  abstract get state(): TransportState;
+  abstract getConfig(): Promise<RTVIClientConfigOption[]>;
+  abstract updateConfig(
+    config: RTVIClientConfigOption[],
+    interrupt?: boolean
+  ): Promise<RTVIMessage>;
+  abstract describeConfig(): Promise<unknown>;
+
+  abstract getServiceOptionsFromConfig(
+    serviceKey: string,
+    config?: RTVIClientConfigOption[]
+  ): Promise<RTVIClientConfigOption | undefined>;
+  abstract getServiceOptionValueFromConfig(
+    serviceKey: string,
+    option: string,
+    config?: RTVIClientConfigOption[]
+  ): Promise<unknown | undefined>;
+  abstract setServiceOptionInConfig(
+    serviceKey: string,
+    option: ConfigOption | ConfigOption[],
+    config?: RTVIClientConfigOption[]
+  ): Promise<RTVIClientConfigOption[] | undefined>;
+  abstract setConfigOptions(
+    configOptions: RTVIClientConfigOption[],
+    config?: RTVIClientConfigOption[]
+  ): Promise<RTVIClientConfigOption[]>;
+
+  abstract action(
+    action: RTVIActionRequestData
+  ): Promise<RTVIMessageActionResponse>;
+  abstract describeActions(): Promise<unknown>;
+  abstract sendMessage(message: RTVIMessage): void;
+  protected abstract handleMessage(ev: RTVIMessage): void;
+
+  abstract registerHelper(
+    service: string,
+    helper: RTVIClientHelper
+  ): RTVIClientHelper;
+  abstract getHelper<T extends RTVIClientHelper>(
+    service: string
+  ): T | undefined;
+  abstract unregisterHelper(service: string): void;
+}
 
 // ----- Deprecated types