+++ rtvi-client-js/src/helpers/index.ts
@@ -1,4 +1,5 @@
-import { Client, RTVIMessage } from "..";
+import { RTVIClient } from "../clients";
+import { RTVIMessage } from "../messages";
 
 export type RTVIClientHelpers = Partial<Record<string, RTVIClientHelper>>;
 
@@ -13,7 +14,7 @@ export interface RTVIClientHelperOptions {
 
 export abstract class RTVIClientHelper {
   protected _options: RTVIClientHelperOptions;
-  protected declare _client: Client;
+  protected declare _client: RTVIClient;
   protected declare _service: string;
 
   constructor(options: RTVIClientHelperOptions) {
@@ -22,7 +23,7 @@ export abstract class RTVIClientHelper {
 
   public abstract handleMessage(ev: RTVIMessage): void;
   public abstract getMessageTypes(): string[];
-  public set client(client: Client) {
+  public set client(client: RTVIClient) {
     this._client = client;
   }
   public set service(service: string) {