+++ rtvi-client-js/src/index.ts
@@ -1,3 +1,4 @@
+export * from "./actions";
 export * from "./clients";
 export * from "./errors";
 export * from "./events";