+++ rtvi-client-js/src/transport/index.ts
@@ -1,3 +1,0 @@
-import { Participant, Tracks, Transport, TransportState } from "./core";
-export { Transport };
-export type { Participant, Tracks, TransportState };